fix: split definitions move test helpers
